// types/ethers-contracts/TodoList.ts
export type TodoItemStruct = {
  todoId: BigNumberish;
  title: string;
  finished: boolean;
  creator: string;
  amount: BigNumberish;
};